refactor(cl): move RestartSentEmailQuotaMixin to call site
RestartSentEmailQuotaMixin is only used by a couple tests in users, so I moved it next to where it is used.  If it gets used elsewhere, then we call pull it out into the common mixins.
