+++ cl/users/tests.py
@@ -56,7 +56,6 @@ from cl.tests.cases import (
     APITestCase,
     ESIndexTestCase,
     LiveServerTestCase,
-    RestartSentEmailQuotaMixin,
     TestCase,
 )
 from cl.tests.mixins import SimpleUserDataMixin
@@ -730,6 +729,7 @@ class DisposableEmailTest(SimpleUserDataMixin, TestCase):
     bad_email = f"{user}@{bad_domain}"
 
     def setUp(self) -> None:
+        super().setUp()
         self.client = AsyncClient()
 
     async def test_can_i_create_account_with_bad_email_address(self) -> None:
@@ -769,6 +769,7 @@ class DisposableEmailTest(SimpleUserDataMixin, TestCase):
 
 class LiveUserTest(BaseSeleniumTest):
     def setUp(self) -> None:
+        super().setUp()
         self.up = UserProfileWithParentsFactory.create(
             user__username="pandora",
             user__password=make_password("password"),
@@ -833,6 +834,7 @@ class LiveUserTest(BaseSeleniumTest):
 class SNSWebhookTest(TestCase):
     @classmethod
     def setUpTestData(cls) -> None:
+        super().setUpTestData()
         test_dir = Path(settings.INSTALL_ROOT) / "cl" / "users" / "test_assets"
         with (
             open(
@@ -1463,6 +1465,22 @@ class SNSWebhookTest(TestCase):
         )
 
 
+class RestartSentEmailQuotaMixin:
+    """Restart sent email quota in redis."""
+
+    @classmethod
+    def restart_sent_email_quota(cls, prefix="email"):
+        r = get_redis_interface("CACHE")
+        keys = r.keys(f"{prefix}:*")
+
+        if keys:
+            r.delete(*keys)
+
+    def tearDown(self):
+        self.restart_sent_email_quota()
+        super().tearDown()
+
+
 @override_settings(
     EMAIL_BACKEND="cl.lib.email_backends.EmailBackend",
     BASE_BACKEND="django.core.mail.backends.locmem.EmailBackend",
@@ -1471,6 +1489,7 @@ class SNSWebhookTest(TestCase):
 class CustomBackendEmailTest(RestartSentEmailQuotaMixin, TestCase):
     @classmethod
     def setUpTestData(cls):
+        super().setUpTestData()
         cls.user = UserFactory()
         test_dir = Path(settings.INSTALL_ROOT) / "cl" / "users" / "test_assets"
         with (
@@ -2505,6 +2524,7 @@ class CustomBackendEmailTest(RestartSentEmailQuotaMixin, TestCase):
 class DeleteOldEmailsTest(TestCase):
     @classmethod
     def setUpTestData(cls):
+        super().setUpTestData()
         cls.msg1 = EmailSentFactory()
         cls.msg2 = EmailSentFactory()
         cls.msg3 = EmailSentFactory()
@@ -2540,6 +2560,7 @@ class DeleteOldEmailsTest(TestCase):
 class RetryFailedEmailTest(RestartSentEmailQuotaMixin, TestCase):
     @classmethod
     def setUpTestData(cls):
+        super().setUpTestData()
         cls.user = UserFactory()
         test_dir = Path(settings.INSTALL_ROOT) / "cl" / "users" / "test_assets"
         with (
@@ -3082,6 +3103,7 @@ class RetryFailedEmailTest(RestartSentEmailQuotaMixin, TestCase):
 class EmailBrokenTest(ESIndexTestCase, TestCase):
     @classmethod
     def setUpTestData(cls):
+        super().setUpTestData()
         test_dir = Path(settings.INSTALL_ROOT) / "cl" / "users" / "test_assets"
         with (
             open(
@@ -3375,16 +3397,19 @@ class WebhooksHTMXTests(APITestCase, TestCase):
 
     @classmethod
     def setUpTestData(cls):
+        super().setUpTestData()
         cls.user_1 = UserFactory()
         cls.user_2 = UserFactory()
 
     def setUp(self) -> None:
+        super().setUp()
         self.webhook_path = reverse("webhooks-list")
         self.client = make_client(self.user_1.pk)
         self.client_2 = make_client(self.user_2.pk)
 
     def tearDown(cls):
         Webhook.objects.all().delete()
+        super().tearDown()
 
     async def make_a_webhook(
         self,
@@ -3813,6 +3838,7 @@ class NeonAccountCreationTest(TestCase):
 @patch("cl.users.views.update_neon_account")
 class NeonAccountUpdateTest(TestCase):
     def setUp(self) -> None:
+        super().setUp()
         self.client = AsyncClient()
         self.up = UserProfileWithParentsFactory.create(
             user__username="pandora",
